edicion de los controladores de mascota
se hicieron cambios en el controlador de mascota y interfaces relacionadas

+++ src/main/java/com/example/ApiVeterinaria/Repository/IMascota.java
@@ -3,5 +3,5 @@ package com.example.ApiVeterinaria.Repository;
 import com.example.ApiVeterinaria.Modelos.Mascota;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IMascota extends CrudRepository<Mascota,Long> {
+public interface IMascota extends CrudRepository<Mascota,Integer> {
 }
